+++ backend/src/main/java/com/fptu/evstation/rental/evrentalsystem/dto/UpdateVehicleDetailsRequest.java
@@ -18,7 +18,7 @@ public class UpdateVehicleDetailsRequest {
     private Double currentMileage;
     @Min(value = 0, message = "Mức pin không thể âm")
     @Max(value = 100, message = "Mức pin không thể lớn hơn 100")
-    private Double batteryLevel;
+    private Integer batteryLevel;
     private VehicleCondition newCondition;
     private VehicleStatus status;
 }
